refactor: refactored NotificationBox
