+++ packages/react-notifications/src/NotificationBox.tsx
@@ -1,12 +1,57 @@
-import { Box } from '@ttoss/ui';
-import { useNotifications } from './Provider';
+import * as React from 'react';
+import { Button, Flex } from '@ttoss/ui';
+import { type NotifyParams, useNotifications } from './Provider';
 
 export const NotificationBox = () => {
-  const { notifications } = useNotifications();
+  const { setNotifications, notifications } = useNotifications();
 
   if (!notifications) {
     return null;
   }
 
-  return <Box>{JSON.stringify(notifications, null, 2)}</Box>;
+  const ButtonMemoized = React.memo(({ message, type }: NotifyParams) => {
+    return (
+      <Button
+        sx={{
+          backgroundColor: type === 'error' ? 'danger' : 'positive',
+        }}
+        onClick={() => {
+          return setNotifications(undefined);
+        }}
+        rightIcon="close"
+        leftIcon={type === 'error' ? 'warning' : undefined}
+      >
+        {message}
+      </Button>
+    );
+  });
+
+  ButtonMemoized.displayName = 'ButtonMemoized';
+
+  return (
+    <Flex
+      sx={{
+        alignItems: 'center',
+        justifyContent: 'center',
+        marginTop:
+          !notifications ||
+          (Array.isArray(notifications) && !notifications.length)
+            ? 0
+            : '2xl',
+      }}
+    >
+      {Array.isArray(notifications) ? (
+        notifications.map((notification) => {
+          return (
+            <ButtonMemoized
+              key={JSON.stringify(notification)}
+              {...notification}
+            />
+          );
+        })
+      ) : (
+        <ButtonMemoized {...notifications} />
+      )}
+    </Flex>
+  );
 };
